+++ packages/wasm-dpp2/src/identity/transitions/mod.rs
@@ -1,0 +1,7 @@
+pub mod create_transition;
+pub mod credit_withdrawal_transition;
+pub mod identity_credit_transfer_transition;
+pub mod masternode_vote_transition;
+pub mod public_key_in_creation;
+pub mod top_up_transition;
+pub mod update_transition;