+++ scripts/release/bump_version.js
@@ -1,8 +1,10 @@
 const fs = require('fs');
 const path = require('path');
 const semver = require('semver');
+const TOML = require('@iarna/toml');
 const packagesIterator = require('../utils/packagesIterator');
 const rootPackageJson = require('../../package.json');
+const rootCargoTomlPath = path.join(__dirname, '..', '..', 'Cargo.toml');
 
 const convertReleaseToPrerelease = (version, prereleaseType) => {
   const bumpedVersion = semver.inc(version, 'minor');
@@ -54,8 +56,8 @@ const parseArgs = (argv) => {
 };
 
 const bumpNpmPackages = (versionFunc, releaseType) => {
-  for (const {filename, json} of packagesIterator.npm()) {
-    const {version} = json;
+  for (const { filename, json } of packagesIterator.npm()) {
+    const { version } = json;
 
     json.version = versionFunc(version, releaseType);
 
@@ -64,18 +66,21 @@ const bumpNpmPackages = (versionFunc, releaseType) => {
 }
 
 const bumpRustPackages = (versionFunc, releaseType) => {
-  for (const {filename, toml} of packagesIterator.rust()) {
-    const {version} = toml.package;
+  const cargoFile = fs.readFileSync(rootCargoTomlPath, 'utf-8');
+  const parsedCargo = TOML.parse(cargoFile);
 
-    const tomlVersion = versionFunc(version, releaseType);
+  const currentVersion = parsedCargo?.workspace?.package?.version;
 
-    const cargoFile = fs.readFileSync(filename, 'utf-8');
-
-    const replaceFrom = `version = "${version}"`;
-    const replaceTo = `version = "${tomlVersion}"`;
-
-    fs.writeFileSync(filename, cargoFile.replace(replaceFrom, replaceTo));
+  if (!currentVersion) {
+    throw new Error('Unable to determine workspace package version from Cargo.toml');
   }
+
+  const nextVersion = versionFunc(currentVersion, releaseType);
+
+  const replaceFrom = `version = "${currentVersion}"`;
+  const replaceTo = `version = "${nextVersion}"`;
+
+  fs.writeFileSync(rootCargoTomlPath, cargoFile.replace(replaceFrom, replaceTo));
 }
 
 (async () => {